searchresults.html
Integrated JINJA2-compatible for loop to display search results. When /searchresults route on the flask backend is invoked, it will use the get_gigs helper function to get the gigs, then those gigs will be passed as an argument as shown below. The gigs will pass into the html-integrated for loop and automatically display in a formatted way. 

@app.route('/searchresults')
def search_results():
    results = get_gigs() 
    return render_template('searchresults.html', results=results)